+++ src/FluentNHibernate.Testing/Visitors/ComponentColumnPrefixVisitorSpecs.cs
@@ -203,6 +203,57 @@ namespace FluentNHibernate.Testing.Visitors
         }
     }
 
+    [TestFixture]
+    public class when_the_component_column_prefix_visitor_processes_a_component_with_a_prefix_using_field_alias : ComponentColumnPrefixVisitorSpec
+    {
+        PersistenceModel model;
+        IEnumerable<HibernateMapping> mappings;
+        ClassMapping targetMapping;
+        const string columnPrefix = "{property}";
+
+        public override void establish_context()
+        {
+            model = new PersistenceModel();
+
+            var componentMap = new ComponentMap<FieldComponent>();
+            componentMap.Map(x => x.X);
+            componentMap.Map(x => x.Y);
+
+            model.Add(componentMap);
+
+            var classMapping = new ClassMap<Root>();
+            classMapping.Id(r => r.Id);
+            classMapping.Component(Reveal.Member<Root, FieldComponent>("component"), cpt => cpt.Access.Field().ColumnPrefix(columnPrefix));
+            model.Add(classMapping);
+        }
+
+        public override void because()
+        {
+            mappings = model.BuildMappings().ToList();
+            targetMapping = mappings.SelectMany(x => x.Classes).FirstOrDefault(x => x.Type == typeof(Root));
+        }
+
+        [Test]
+        public void should_prefix_field_columns()
+        {
+            targetMapping.Components.Single()
+                .Properties.SelectMany(x => x.Columns)
+                .Each(c => c.Name.ShouldStartWith("component"));
+        }
+    }
+
+    class Root
+    {
+        FieldComponent component;
+        public int Id { get; set; }
+    }
+
+    class FieldComponent
+    {
+        public string X { get; set; }
+        public int? Y { get; set; }
+    }
+
     public abstract class ComponentColumnPrefixVisitorSpec : Specification
     {
         protected AnyMapping any_with_column(string column)
